+++ client/src/index.js
@@ -1,8 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import {BrowserRouter} from 'react-router-dom';
+
 import './index.css';
 import Application from './boot/Application';
-import registerServiceWorker from './registerServiceWorker';
+import register from './register';
 
-ReactDOM.render(<Application/>, document.getElementById('root'));
-registerServiceWorker();
+ReactDOM.render((
+    <BrowserRouter>
+        <Application/>
+    </BrowserRouter>
+), document.getElementById('root'));
+register();
